data_dispatcher-service/requirements.txt
alembic==1.11.1
apifairy==1.4.0
apispec==6.4.0
attrs==23.2.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
confluent-kafka==2.2.0
decorator==5.1.1
flasgger==0.9.7b2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.1.0
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.6
influxdb-client==1.36.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
mistune==3.0.2
mysqlclient==2.2.0
numpy==1.26.3
packaging==23.1
pandas==2.1.4
py==1.11.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3.post1
PyYAML==6.0.1
reactivex==4.0.4
referencing==0.32.1
requests==2.31.0
retry==0.9.2
rpds-py==0.17.1
simple-websocket==0.10.1
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.0.4
webargs==8.4.0
Werkzeug==2.3.6
wsproto==1.2.0
zope.event==5.0
zope.interface==6.0